Moving cpu_info to benchmark_util. Only importing permutation_reps in python.
